Feat/update game (#65)
* feat: unserialize data to replace with server info

* feat: serialize every networkable components

* fix: remove uint8 from collider as it's handled as char by the serializer

* add a network identifier to the players

* feat: retrieve input from players

* feat: update players according to server

* fix: give a NetworkComponent to each client

the component is given only to the player assigned to the client

* fix: put shoot logic in server

* feat: add color to sprite

* fix: check if entity has component

* change: add a clock to the server to send data

* feat: add player shoot to network

* change: send only other player inputs to a player

* fix: handle shoot with key up instead of down

* refacto: remove debug

---------

+++ common/include/components/Projectile.hpp
@@ -12,11 +12,12 @@
 
 namespace component
 {
-    enum class ProjectileType : uint8_t
+    enum class ProjectileType : int
     {
         PLAYER = 0,
         MOB = 1
     };
+
     struct Projectile
     {
         ProjectileType type;
